Completed DAOCompteur, JDBCCompteur

See merge request pxn4923a/saea3.01!7

+++ DiagrammeSAE3/src/modele/BienLouable.java
@@ -22,7 +22,7 @@ public class BienLouable extends BienImmobilier {
 		this.compteur = compteur;
 	}
 
-	public void changerCompteur(double indexActuel) {
+	public void changerCompteur(int indexActuel) {
 		this.compteur.setIndexActuel(indexActuel);
 	}
